+++ convert.js
@@ -4,6 +4,7 @@ const fileArg = require("./src/file-arg");
 const fileCheck = require("./src/file-check");
 const fileRead = require("./src/file-read");
 const procData = require("./src/proc-data");
+const outputFolder = require("./src/output-folder");
 
 runDataNormalization();
 
@@ -13,8 +14,11 @@ function runDataNormalization()
 	var givenFileType = fileArg.readFileType();
 	var fileTypeFlag = fileArg.prepareFileType(givenFileType);
 	
-	clear();
-	callInputFileCheck(fileTypeFlag);
+	if (fileTypeFlag >= 0)
+	{
+		clear();
+		callInputFileCheck(fileTypeFlag);
+	}
 }
 
 
@@ -24,7 +28,7 @@ function callInputFileCheck(conversionTypeFlag)
 	{
 		if (inpChkErr !== null)
 		{
-			exitProgram.callError(inpChkErr.message);
+			exitProgram.callError(inpChkErr.message, true);
 		}
 		else
 		{
@@ -40,7 +44,7 @@ function callInputFileRead(convTypeFlag)
 	{
 		if (inpReadErr !== null)
 		{
-			exitProgram.callError(inpReadErr.message);
+			exitProgram.callError(inpReadErr.message, true);
 		}
 		else
 		{
@@ -52,19 +56,48 @@ function callInputFileRead(convTypeFlag)
 
 function callDataProcessing(retrievedCsvData, convType)
 {
-	procData.processData(retrievedCsvData, function (dataProcErr, dataProcRes)
+	procData.processData(retrievedCsvData, function (dataProcErr, normalizedDataObject)
 	{
 		if (dataProcErr !== null)
 		{
-			exitProgram.callError(dataProcErr.message);
+			exitProgram.callError(dataProcErr.message, true);
 		}
 		else
 		{
-			console.log("");
-			console.log("Orders: " + dataProcRes.orderEntries.length);
-			console.log("Items: " + dataProcRes.orderItems.length);
-			console.log("");
-			console.log(dataProcRes.orderItems[0]);
+			callOutputFolder(normalizedDataObject, convType);
 		}
 	});
+}
+
+
+function callOutputFolder(normDataObj, cTypeFlag)
+{
+	outputFolder.createFolder(function (outputFolderErr, outputFolderPath)
+	{
+		if (outputFolderErr !== null)
+		{
+			exitProgram.callError(outputFolderErr.message, true);
+		}
+		else
+		{
+			callOutputFileWrite(normDataObj, outputFolderPath, cTypeFlag);
+		}
+	});
+}
+
+
+function callOutputFileWrite(normData, oFolderPth, cType)
+{
+	if (cType > 0)
+	{
+		exitProgram.callError("SQL Definition", true);
+	}
+	else if (cType === 0)
+	{
+		exitProgram.callError("CSV Data", true);
+	}
+	else
+	{
+		exitProgram.callError("Invalid", true);
+	}
 }